requirements.txt
Django~=5.1.3
celery~=5.4.0
redis~=5.2.0
requests~=2.32.3
python-dotenv~=1.0.1
python-telegram-bot~=21.7
asgiref~=3.8.1
dash~=2.18.2
channels~=3.0.5
bokeh~=3.6.1